Move VisitInfo, ParticipantVisits classes
