+++ common/src/python/submissions/BUILD
@@ -1,0 +1,1 @@
+python_sources(name="lib")
